backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
asyncpg
alembic
pydantic-settings
redis
celery
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests
email-validator
bcrypt==4.0.1
psycopg2-binary
pytest
httpx
pytest-asyncio
